feat(remove-acf): removed Advanced Custom Fields as a bundled dependency
Since ACF is not required for the Hoverboard parent theme to operate,
it's been removed as a bundled dependency in favor of requiring it from
the child theme.

BREAKING CHANGE: Any child themes using Advanced Custom Fields need to
be updated to the latest version of hoverboard-child OR need to have
ACF installed the usual WordPress way to avoid a broken theme.